+++ go-backend/model/node.go
@@ -19,6 +19,7 @@ type Node struct {
 	CreatedTime int64  `gorm:"column:created_time" json:"createdTime"`
 	UpdatedTime int64  `gorm:"column:updated_time" json:"updatedTime"`
 	Status      int    `gorm:"column:status" json:"status"`
+	Inx         int    `gorm:"column:inx" json:"inx"`
 }
 
 func (Node) TableName() string {