+++ go-backend/model/tunnel.go
@@ -17,6 +17,7 @@ type Tunnel struct {
 	CreatedTime    int64   `gorm:"column:created_time" json:"createdTime"`
 	UpdatedTime    int64   `gorm:"column:updated_time" json:"updatedTime"`
 	Status         int     `gorm:"column:status" json:"status"`
+	Inx            int     `gorm:"column:inx" json:"inx"`
 }
 
 func (Tunnel) TableName() string {